+++ src/options.rs
@@ -152,4 +152,7 @@ pub struct Options {
     pub jsonrpc_omit_jsonrpc: bool,
     pub inhibit_pongs: Option<usize>,
     pub max_sent_pings: Option<usize>,
+
+    pub lengthprefixed_header_bytes: usize,
+    pub lengthprefixed_little_endian: bool,
 }